Finished converting to spring-data-rest.
Removed initial CRUD controller for case management systems, which is now
irrelevant, and updated the configuration to allow complete hiding of internal
IDs for case management systems and types (sadly not readily possible for
the cases themselves).

Also tweaked the data model to remove the illogical and unmaintainable
one-to-many relationship between case managers and case types.