Added somewhat robust service testing.
Added tests for the CsvLoader service class, which incidentally exercise
a certain amount of other service code, and added test infrastructure to
allow running tests that modify the database in any order (by wiping the
DB before each test).

A certain amount of refactoring and tweaking was involved, some of which
is just general and some of which was needed to make this all pass
Checkstyle.

Ratcheted the test-coverage threshold up to 48%, that being the level
that this now brings it to.