Made handling of excessive in-list arguments more consistent.
In HSQLDB, we can have an arbitrarily large in-list in a query, but in
Postgresql it tops out at roughly 2**16-1 (because of a JDBC issue, we
think). Added validation to the repository so it will consistently fail
at 30,000 entries, which should really be enough to get us by for a while.