Set up Spring Boot Actuator, and adjusted permissions.
Explicitly configured actuators other than the default "health"
endpoint, and made them inaccessible other than to users with an
appropriately isolated role.

Added build and git-status information to the "info" actuator,
and made that one (only) accessible to anonymous users, so that
it can be used to verify deployments.

As a throw-in, added anonymous access to /favicon.ico, because
seriously that should be automatic.