Fixed raw-SQL migrations to work with all schemas/usernames.
The original form only works if the default schema for the migrations user
is the same as the default-schema setting for liquibase, which is not always
true. Added a shim for that case, and a second shim to get around the
integrity check that was violated by the first shim.