Added auth-debugging controller to help with x509 development.
Added back in the old authentication-dumping behavior, on a profile
that is no longer required for the actual basic functioning of the
application, to ease exploration of the x509 authentication data
structures.

Added a minor Jackson configuration tweak to make this controller actually
work with the data that Spring Security throws at it when we do this.