Updated database and code to allow longer user IDs.
Made the Spring Session persistence layer use the same maximum length
for user IDs that is used by our internal data structures, and updated
the X509 mapping code (where this limit is most likely to be encountered)
to trap and fail on that error condition.

Added some tests that depend on files that are not yet checked in, because
too many manual steps are involved and some automation streamlining seems
needful.