Added mapping of permissions for x509 users.
Created property-mapping and translation infrastructure to allow specifying
permissions for individual pre-authenticated users with known distinguished
names. (This is not very flexible, but is adequate to known current needs.)

Also made the thing a little more flexible than is strictly necessary, to
try to leave space for unifying the three different ways we currently allow
users and authorizations to be specified.