Dependency patch update.
Updated Spring Boot to the latest release on this branch (2.1.14), removed
the now-unnecessary override for jackson-databind, and added one for tomcat.

Since the tomcat override is best done using external properties, removed the
inlined externalized properties (which was always a little silly) in favor
of a standard gradle.properties file.