De-protected an entity constructor for Hibernate.
Probably to do with the ongoing Javassist/Bytebuddy wars: previously all
entity constructors needed to exist but could be private, but for Boot 2.2,
this one (only) needs to change to be protected. It appears that this can
be reverted when we upgrade to 2.3, and no, I have no idea why.