Configuration and testing changes for changed static layout.
We have a different set of URLs for both the API browsers, and no default favicon
any more: updated security config to allow for this, added a courtesy redirect,
and tweaked the redirect-preventing code to allow that specific one to work.
Tests are of course updated to expect the now-correct behavior.